+++ tests/message_test.py
@@ -163,6 +163,23 @@ class EvilObject(object):
     pass
 
 
+class EncodedTest(testlib.TestCase):
+    klass = mitogen.core.Message
+    def test_ctor(self):
+        msg = self.klass.encoded(42, self.klass.ENC_PKL)
+        self.assertEqual(self.klass.ENC_PKL, msg.enc)
+
+        msg = self.klass.encoded(b'abc', self.klass.ENC_BIN)
+        self.assertEqual(b'abc', msg.data)
+        self.assertEqual(self.klass.ENC_BIN, msg.enc)
+
+    def test_invalid_args(self):
+        self.assertRaises(ValueError, lambda: self.klass.encoded(42, enc=self.klass.ENC_MGC))
+        self.assertRaises(ValueError, lambda: self.klass.encoded(b('abc'), enc=self.klass.ENC_MGC))
+        self.assertRaises(Exception, lambda: self.klass.encoded(42, enc=self.klass.ENC_BIN))
+        self.assertRaises(Exception, lambda: self.klass.encoded(u'abc', enc=self.klass.ENC_BIN))
+
+
 class PickledTest(testlib.TestCase):
     # getting_started.html#rpc-serialization-rules
     klass = mitogen.core.Message
@@ -368,6 +385,11 @@ class UnpickleTest(testlib.TestCase):
     def test_no_throw_dead(self):
         m = self.klass.pickled('derp', reply_to=mitogen.core.IS_DEAD)
         self.assertEqual('derp', m.unpickle(throw_dead=False))
+
+    def test_invalid_enc(self):
+        msg = self.klass.pickled(42)
+        msg.enc = self.klass.ENC_BIN
+        self.assertRaises(ValueError, msg.unpickle)
 
 
 class UnpickleCompatTest(testlib.TestCase):