+++ docs/changelog.rst
@@ -26,6 +26,8 @@ In progress (unreleased)
 * :gh:issue:`1430` :mod:`mitogen`: Explicitly mark messages known to carry
   pickled data, using :data:`mitogen.core.Message.ENC_PKL`. This repurposes
   the magic field as a content encoding enumeration.
+* :gh:issue:`1430` :mod:`mitogen`: Add explicit binary Message encoding,
+  marked using :data:`mitogen.core.Message.ENC_BIN`.
 
 
 v0.3.38 (2026-01-23)